+++ src/commons/input/tensors.rs
@@ -1,3 +1,5 @@
+use std::collections::HashMap;
+
 use ort::session::SessionInputs;
 use crate::util::result::Result;
 use crate::commons::input::encoded::EncodedInput;
@@ -5,6 +7,7 @@ use crate::commons::input::encoded::EncodedInput;
 const TENSOR_INPUT_IDS: &str = "input_ids";
 const TENSOR_ATTN_MASKS: &str = "attention_mask";
 const TENSOR_TOKEN_TYPE_IDS: &str = "token_type_ids";
+const TENSOR_POSITION_IDS: &str = "position_ids";
 
 
 /// Input tensors, ready for inferences
@@ -15,17 +18,23 @@ pub struct InputTensors<'a> {
 
 impl<'a> InputTensors<'a> {
     fn make_session_inputs(input: EncodedInput) -> Result<SessionInputs<'a, 'a>> {
+        let mut inputs: HashMap<&str, ort::value::Value<ort::value::TensorValueType<i64>>> = HashMap::new();
+        inputs.insert(TENSOR_INPUT_IDS, ort::value::Value::from_array(input.input_ids)?);
+        inputs.insert(TENSOR_ATTN_MASKS, ort::value::Value::from_array(input.attn_masks)?);
         if let Some(type_ids) = input.type_ids {
-            Ok(ort::inputs! { TENSOR_INPUT_IDS => input.input_ids, TENSOR_ATTN_MASKS => input.attn_masks, TENSOR_TOKEN_TYPE_IDS => type_ids }?.into())
+            inputs.insert(TENSOR_TOKEN_TYPE_IDS, ort::value::Value::from_array(type_ids)?);
         }
-        else {
-            Ok(ort::inputs! { TENSOR_INPUT_IDS => input.input_ids, TENSOR_ATTN_MASKS => input.attn_masks }?.into())
-        }        
+        if let Some(position_ids) = input.position_ids {
+            inputs.insert(TENSOR_POSITION_IDS, ort::value::Value::from_array(position_ids)?);
+        } 
+        Ok(SessionInputs::try_from(inputs)?)        
     }
 
-    pub fn input_tensors(with_token_types: bool) -> Vec<&'static str> {
-        if with_token_types { vec![TENSOR_INPUT_IDS, TENSOR_ATTN_MASKS, TENSOR_TOKEN_TYPE_IDS] }
-        else { vec![TENSOR_INPUT_IDS, TENSOR_ATTN_MASKS] }
+    pub fn input_tensors(with_token_types: bool, with_position_ids: bool) -> Vec<&'static str> {
+        let mut result = vec![TENSOR_INPUT_IDS, TENSOR_ATTN_MASKS];
+        if with_token_types { result.push(TENSOR_TOKEN_TYPE_IDS); }
+        if with_position_ids { result.push(TENSOR_POSITION_IDS); }
+        result
     }
 }